+++ include/dirichletBoundaryCreator.hpp
@@ -13,7 +13,7 @@ template<int DIM>
 class DirichletBoundaryCreator: public BoundaryCreatorInterface{
     public:
             //CTOR
-            DirichletBoundaryCreator(float* _ptr2State, int _Nx, int _Ny, int _Nz=0, float _surfaceTemp=0.): BoundaryCreatorInterface(_ptr2State, _Nx, _Ny, _Nz, _surfaceTemp){}
+            DirichletBoundaryCreator(double* _ptr2State, int _Nx, int _Ny, int _Nz=0, double _surfaceTemp=0.): BoundaryCreatorInterface(_ptr2State, _Nx, _Ny, _Nz, _surfaceTemp){}
             BoundaryConditionTypeInterface* createBoundaryType() override{return(new DirichletImpl<DIM>());}
 
 };
